requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pydantic
python-dotenv
groq
pypdf
pandas
numpy


